Integrated pdf.js

+++ src/ppshw/fileupload/FileManager.js
@@ -7,11 +7,13 @@ var fs=require('fs')
   , crypto=require('crypto')
   , mongoose=require('mongoose')
   , File=mongoose.model('File')
+  , officeRunner=__dirname+'/OfficeWrapper.js'
+  , pdfjsRunner=__dirname+'/pdf_jsWrapper.js'
 ;
 
 var FileManager = {
     init : function(){
-      this.child = require('child_process').fork(__dirname+"/OfficeWrapper.js");
+      this.child = require('child_process').fork(__dirname+"/childprocess.js");
       this.taskId = 0;
       this.tasks = {};
       this.child.on('message', function(message) {
@@ -24,30 +26,40 @@ var FileManager = {
     storeFile: function(filepath,digest,ext,filename){
       
       File.findOne({'digest':digest},function(err, file){
-        if(file===null){
+        if(file!==null){
 //          file.creationDate=Date.now;
 //          file.filename.full=filename;
 //          file.save();
-//        }else{
+        }else{
           file=new File({digest:digest, filename:{full:filename}, tags:'doc'});
           file.save();
-          FileManager.doStoreFile(filepath,digest,ext);
         }
+        FileManager.doStoreFile(filepath,digest,ext);
       });
     },
     doStoreFile: function(filepath,digest,ext){
       var new_filename=digest;
       var storage_path=ppshw.system.Config.get('ppshw:application:upload:dir');
       var content_path=storage_path + digest + '_cnt';
-      var new_path=content_path+'\\' + new_filename+'.'+ext;
+      var new_path=content_path + '/' + new_filename+'.'+ext;
       
-      fs.mkdir(content_path);
+      if(!fs.existsSync(new_path)){
       
-      fs.rename(filepath,new_path);
-      if(this.isOfficeFile(ext)){
+        fs.mkdir(content_path);
+      
+        fs.rename(filepath,new_path);
+        if(this.isOfficeFile(ext)){
         //let's create a pdf
-        this.createPdf(content_path,new_path,new_filename);
-      };
+          this.createPdf(content_path,new_path,new_filename);
+          this.createPreview(content_path);
+        }
+      }
+    },
+    createPreview : function(content_path){
+      
+    },
+    isPdfFile : function(ext){
+      return ext==='pdf';
     },
     isOfficeFile : function(ext){
       switch(ext){
@@ -60,9 +72,13 @@ var FileManager = {
       case 'ppt':
       case 'pptx':
       case 'odf':
+      case 'otf':
       case 'odt':
-      case 'odc':
       case 'ott':
+      case 'odc':
+      case 'otc':
+      case 'odp':
+      case 'otp':
         return true;
       default:
           return false;
@@ -78,12 +94,13 @@ var FileManager = {
     },
     createPdf : function(content_path,filepath,filename){
       var cmd=this.getConvertCommand(filepath,content_path);
-      this.addTask(cmd);
+      this.addTask(cmd,officeRunner);
     },
-    addTask : function(data, callback) {
+    addTask : function(data, runner, callback) {
+      console.log(runner);
       var id = this.taskId++;
       this.tasks[id] = callback;
-      this.child.send({id: id, data: data});
+      this.child.send({id: id, data: data, runner: runner});
     }
 };
 FileManager.init();
@@ -93,7 +110,7 @@ exports.handleUpload = function(filename, file){
   var ext = filename.split('.');
   ext=ext[ext.length-1];
   
-  var hash = crypto.createHash('md5')
+  var hash = crypto.createHash('sha256')
     , stream = fs.createReadStream(file.path, {encoding: 'binary'})
   ;
